Adds solutions for beginner levels 4-6.
Kinda stoked with how these turned out, even though lvl6 turned out to be piles
more code. 4 to 5 was a single line, which makes me feel better about my
approach. In lvl6, the `wat_do` method changed return types to an array so that
I can more cleanly control when to include a direction with the return.